Avoid an extra shell process before launching tail
Before this commit:

  508 $ pstree -sp $$
  init(1)───tmux(13543)───bash(13546)───pstree(13646)
  509 $ pstree -p 13543
  tmux(13543)─┬─bash(13544)───tail(13637)
              └─bash(13546)───pstree(13647)

After this commit:

  508 $ pstree -sp $$
  init(1)───tmux(13718)───bash(13721)───pstree(13778)
  509 $ pstree -p 13718
  tmux(13718)─┬─bash(13721)───pstree(13779)
              └─tail(13719)

The extra shell (13544 above) made exiting tedious, since ^C to kill
the tail (13637 above) would drop you into the shell, and you'd need
another 'exit' or ^D to close the pane.

It seems like split-window always breaks the target (defaults to
current) pane, keeping the old process running in the top pane and
launching the new process in the bottom pane.  We could use
'swap-pane', but it seems easier to put the tail in the initial pane
and the shell in the second pane.

I've also dropped the -n from tail, since we're now launching the tail
right after truncating the log file.