Clear tmux's scrollback buffer
This is a bit awkward, because we don't have a good way to block until
Bash acts on the C-l.  POSIX only requires support for integer seconds
[1], but Linux distributions will mostly be using the GNU Coreutils
implementation which supports fractional seconds [2], and OS X [3] and
FreeBSD [4] also support fractional seconds.  MSysGit doesn't seem to
document their choice, but the binary [5] seems to be from Shellutils
[6], which has since been folded into Coreutils:

  $ wget https://github.com/msysgit/msysgit/blob/master/bin/sleep.exe
  $ strings sleep.exe | grep -i 'gnu\|copyright'
  Report bugs to <bug-sh-utils@gnu.org>.
  GNU sh-utils
  Copyright (C) 1999 Free Software Foundation, Inc.

Whether or not that old version actually supports fractional seconds
or not is unclear.  The functionality landed in 1999 [7], but it was
released in 2.0a [8], which was cut in 2000 [9,10].  So I'd guess the
MSysGit version doesn't support fractional seconds, but waiting a
whole second is way to slow.  If the local sleep doesn't support
fractional seconds, the consequences would likely just be an error
message on stderr and not clearing a few commands from the scrollback
buffer.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/sleep.html
[2]: http://www.gnu.org/software/coreutils/manual/html_node/sleep-invocation.html#sleep-invocation
[3]: https://developer.apple.com/library/mac/documentation/Darwin/Reference/ManPages/man1/sleep.1.html
[4]: https://www.freebsd.org/cgi/man.cgi?query=sleep%281%29
[5]: https://github.com/msysgit/msysgit/blob/master/bin/sleep.exe
[6]: http://www.gnu.org/software/shellutils/shellutils.html
[7]: http://git.savannah.gnu.org/gitweb/?p=coreutils.git;a=commit;h=56172516598e1f48a629830aa5ce2c9ab9240bb7
[8]: http://lists.gnu.org/archive/html/bug-sh-utils/2001-11/msg00024.html
[9]: http://git.savannah.gnu.org/gitweb/?p=coreutils.git;a=tag;h=1a06c1bf4e3bb4b6a7958a853e5bf7dba686edd7
[10]: http://git.savannah.gnu.org/gitweb/?p=coreutils.git;a=commit;h=60ff88e1103a998023e3e6febc2a92c2f3341b07